LayoutTests/http/tests/security/blob-null-url-location-origin-expected.txt

CONSOLE MESSAGE: line 1: data URL frame loaded
CONSOLE MESSAGE: line 1: blob popup opened
CONSOLE MESSAGE: line 1: blob popup loadednull
CONSOLE MESSAGE: line 1: blob popup message posted
PASS
